Java sprint 11 v1.1
переделаны зависимости
добавлены классы в пакете error
заменены сигнатуры в User Film UserService FilmService

+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -9,7 +9,7 @@ public interface FilmStorage {
 
     public Film updateFilm(Film film);
 
-    public Film getFilm(int id);
+    public Film getFilm(long id);
 
     public List<Film> getFilms();
 }
